+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/ReserveController.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.presentation;
+package com.wanted.preonboarding.ticket.presentation.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.wanted.preonboarding.ticket.application.TicketSeller;
